+++ src/phanos/messaging.py
@@ -4,13 +4,16 @@ import typing
 
 import aio_pika
 import orjson
-import pika
 import pika.exceptions
+from aio_pika import Message
+from aio_pika.abc import AbstractRobustConnection, AbstractRobustChannel, AbstractRobustExchange
+from pika import ConnectionParameters
 from pika.adapters.utils.connection_workflow import AMQPConnectorException
+from pika.credentials import PlainCredentials
 
 from . import types
 
-__all__ = ("LoggerLike", "NETWORK_ERRORS", "BlockingPublisher")
+__all__ = ("LoggerLike", "NETWORK_ERRORS", "AsyncioPublisher")
 
 LoggerLike = typing.Union[logging.Logger, logging.LoggerAdapter]
 
@@ -22,7 +25,7 @@ NETWORK_ERRORS = (
 )
 
 
-class BlockingPublisher:
+class AsyncioPublisher:
     """Simple blocking AMQP consumer"""
 
     __slots__ = (
@@ -31,14 +34,16 @@ class BlockingPublisher:
         "connection_parameters",
         "connection",
         "channel",
+        "exchange",
         "retry",
         "logger",
     )
     exchange_name: str
     exchange_type: str
-    connection_parameters: pika.ConnectionParameters
-    connection: typing.Optional[pika.BlockingConnection]
-    channel: typing.Optional[pika.adapters.blocking_connection.BlockingChannel]
+    connection_parameters: ConnectionParameters
+    connection: typing.Optional[AbstractRobustConnection]
+    channel: typing.Optional[AbstractRobustChannel]
+    exchange: typing.Optional[AbstractRobustExchange]
     retry: int
     logger: LoggerLike
 
@@ -83,13 +88,14 @@ class BlockingPublisher:
         self.logger = logger or logging.getLogger(__name__)
         self.connection = None
         self.channel = None
+        self.exchange = None
         self.exchange_name = exchange_name
         self.exchange_type = exchange_type.value if isinstance(exchange_type, aio_pika.ExchangeType) else exchange_type
         self.retry = max(0, retry)
-        self.connection_parameters = pika.ConnectionParameters(
+        self.connection_parameters = ConnectionParameters(
             host=host,
             port=port,
-            credentials=pika.credentials.PlainCredentials(
+            credentials=PlainCredentials(
                 username=user or "guest",
                 password=password or "guest",
             ),
@@ -103,46 +109,55 @@ class BlockingPublisher:
         return self.is_connected() and self._is_bound()
 
     def is_connected(self) -> bool:
-        return not (self.connection is None or not self.connection.is_open)
+        return not (self.connection is None or self.connection.is_closed)
 
     def _is_bound(self) -> bool:
-        return not (self.channel is None or not self.channel.is_open)
+        return not (self.channel is None or self.channel.is_closed)
 
-    def close(self) -> None:
+    async def close(self) -> None:
         if self.is_connected():
             if self._is_bound():
-                self.channel.close()
-            self.connection.close()
+                await self.channel.close()
+            await self.connection.close()
         self.connection = None
         self.channel = None
         self.logger.info(f"{type(self).__qualname__} - closed connection")
 
-    def connect(self) -> None:
-        self.connection = pika.BlockingConnection(self.connection_parameters)
-        self.channel = self.connection.channel()
-        self.channel.exchange_declare(
-            exchange=self.exchange_name,
-            exchange_type=self.exchange_type,
+    async def connect(self) -> None:
+        self.connection = await aio_pika.connect_robust(
+            host=self.connection_parameters.host,
+            port=self.connection_parameters.port,
+            login=self.connection_parameters.credentials.username,
+            password=self.connection_parameters.credentials.password,
+            heartbeat=self.connection_parameters.heartbeat,
+            connection_attempts=self.connection_parameters.connection_attempts,
+            retry_delay=self.connection_parameters.retry_delay,
+            blocked_connection_timeout=self.connection_parameters.blocked_connection_timeout,
+        )
+        self.channel = await self.connection.channel()
+        await self.channel.declare_exchange(
+            name=self.exchange_name,
+            type=self.exchange_type,
         )
         self.logger.info(f"{type(self).__qualname__} - connection established")
 
-    def reconnect(self, silent: bool = False) -> None:
+    async def reconnect(self, silent: bool = False) -> None:
         """Force reconnect to server"""
-        self.close()
+        await self.close()
         if not silent:
-            self.connect()
+            await self.connect()
             return
         try:
-            self.connect()
+            await self.connect()
         except NETWORK_ERRORS as e:
             self.logger.warning(f"{type(self).__qualname__} - connection failed on {e!r}")
 
-    def check_or_rebound(self) -> None:
+    async def check_or_rebound(self) -> None:
         if not self:
-            self.close()
-            self.connect()
+            await self.close()
+            await self.connect()
 
-    def publish(self, record: types.Record) -> bool:
+    async def publish(self, record: types.Record) -> bool:
         """
         Push record to message queue.
         :param record: record dict
@@ -151,13 +166,15 @@ class BlockingPublisher:
         attempts_left = 1 + self.retry
         bin_message = orjson.dumps(record)
         is_published = False
-        self.check_or_rebound()
+        await self.check_or_rebound()
         while attempts_left > 0 and not is_published:
             attempts_left -= 1
             try:
-                self.channel.basic_publish(
-                    exchange=self.exchange_name,
-                    body=bin_message,
+                await self.exchange.publish(
+                    message=Message(
+                        body=bin_message,
+                        content_type="application/json",
+                    ),
                     routing_key=record["job"],
                 )
                 is_published = True
@@ -167,5 +184,5 @@ class BlockingPublisher:
                     f"accept message {record!r} because {e!r}"
                 )
                 time.sleep(float(self.connection_parameters.retry_delay))
-                self.reconnect(silent=True)
+                await self.reconnect(silent=True)
         return is_published